requirements.txt
pillow~=8.1.2
torch~=1.8.0
torchvision~=0.9.0
fastapi~=0.115.12